+++ lib/components/auth_gate.dart
@@ -4,7 +4,6 @@ import 'package:firebase_ui_oauth_google/firebase_ui_oauth_google.dart';
 import 'package:flutter/material.dart';
 
 class AuthGate extends StatelessWidget {
-
   final Widget destinationWidget;
 
   const AuthGate({Key? key, required this.destinationWidget}) : super(key: key);
@@ -16,35 +15,33 @@ class AuthGate extends StatelessWidget {
       builder: (context, snapshot) {
         if (!snapshot.hasData) {
           return SignInScreen(
-            providers: [
-              EmailAuthProvider(),
-              GoogleProvider(clientId: "259507610442-e9eq4ogqdvgl19s24hnibsf5i30tlug6.apps.googleusercontent.com")
-            ],
+            providers: [EmailAuthProvider(), GoogleProvider(clientId: "")],
             headerBuilder: (context, constraints, shrinkOffset) {
               return const Padding(
-                padding: EdgeInsets.all(20),
-                child: AspectRatio(
-                  aspectRatio: 1,
-                  child: Image(image: AssetImage('images/CampHarmonyLogo.jpg')),
-                )
-              );
-              },
+                  padding: EdgeInsets.all(20),
+                  child: AspectRatio(
+                    aspectRatio: 1,
+                    child:
+                        Image(image: AssetImage('images/CampHarmonyLogo.jpg')),
+                  ));
+            },
             subtitleBuilder: (context, action) {
               return Padding(
                 padding: const EdgeInsets.symmetric(vertical: 8.0),
                 child: action == AuthAction.signIn
-                    ? const Text('Welcome to Camp Harmony! Please sign in before'
-                    ' checking in')
+                    ? const Text(
+                        'Welcome to Camp Harmony! Please sign in before'
+                        ' checking in')
                     : const Text('Welcome to Camp Harmony! Please register '
-                    'before checking in'),
+                        'before checking in'),
               );
-              },
+            },
             footerBuilder: (context, action) {
               return const Padding(
                 padding: EdgeInsets.only(top: 16),
                 child: Text(
                   'Once signed in you\'ll be able to access personalized '
-                      'resources, like your schedule',
+                  'resources, like your schedule',
                   style: TextStyle(color: Colors.grey),
                   textAlign: TextAlign.center,
                 ),
@@ -55,9 +52,9 @@ class AuthGate extends StatelessWidget {
                   padding: EdgeInsets.all(20),
                   child: AspectRatio(
                     aspectRatio: 1,
-                    child: Image(image: AssetImage('images/CampHarmonyLogo.jpg')),
-                  )
-              );
+                    child:
+                        Image(image: AssetImage('images/CampHarmonyLogo.jpg')),
+                  ));
             },
           );
         }
