+++ lib/components/camp_map.dart
